% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_style.R
\name{apa_style}
\alias{apa_style}
\title{apa_style}
\usage{
apa_style(tab, font = "times", title = "Table 1", subtitle = "Table Title")
}
\arguments{
\item{tab}{gt_tbl A gt table to apply style to.}

\item{font}{string A string with an installed font.}

\item{title}{string A title  (e.g., Table 1).}

\item{subtitle}{string A subtitle (e.g., Descriptives).}
}
\value{
gt_tbl
}
\description{
An opinionated style to apply to tables that is APA 7th compliant.
}
\details{
\itemize{
\item For fonts beyond those installed in base R use \href{https://cran.r-project.org/web/packages/extrafont/README.html}{extrafont package}
\item Titles should be enumerated e.g. Table 1
\item Subtitles should be brief and in sentence case
}
}
\examples{
\dontrun{gt(iris) \%>\% apa_style()}
}
\keyword{style}
