% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_to_pdf.R
\name{check_docto_installed}
\alias{check_docto_installed}
\title{Check \code{docto} is installed}
\usage{
check_docto_installed()
}
\value{
Nothing. Prints error message if not installed.
}
\description{
Check \code{docto} is installed
}
\keyword{internal}
