% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{create_todo}
\alias{create_todo}
\title{create_todo}
\usage{
create_todo(rmd, path = NULL)
}
\arguments{
\item{rmd}{path to rmarkdown file to extract html comments from}

\item{path}{output path. If null (default) the todo list is printed to console}
}
\description{
Creates a todo list
}
\details{
Reads all html comments from a rmarkdown file. Those with |TODO| will be extracted and sorted by assignment. Use @{name} e.g. '@james' to assign a task to someone.
}
\examples{
library(tidyMB)
filepath <- system.file("tests/test_doc.rmd", package = "tidyMB")
create_todo(filepath)
}
