% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_to_pdf.R
\name{docx_to_pdf}
\alias{docx_to_pdf}
\title{docx_to_pdf}
\usage{
docx_to_pdf(path, out)
}
\arguments{
\item{path}{input path}

\item{out}{output path}
}
\description{
Convert docx to PDF
}
\details{
requires a docto installation to work
}
