% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{extract_comments}
\alias{extract_comments}
\title{extract_comments}
\usage{
extract_comments(path)
}
\arguments{
\item{path}{path to file to read}
}
\description{
Reads in a file and pulls out text between comments (\<!--{text}--\>)
}
