% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_functions.R
\name{extract_md_sections}
\alias{extract_md_sections}
\title{extract_sections
Extracts markdown sections}
\usage{
extract_md_sections(path)
}
\description{
extract_sections
Extracts markdown sections
}
