% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_functions.R
\name{get_revision}
\alias{get_revision}
\title{get_revision}
\usage{
get_revision(manuscript, id, quote = TRUE, evaluate = TRUE)
}
\arguments{
\item{id}{the id from a html tag}

\item{quote}{is the output chunk quoted?}

\item{evaluate}{bool. Should inline rchunks be executed?}

\item{path}{the path to an rmarkdown manuscript}
}
\description{
Extract and format revision
}
