% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_metadata.R
\name{grab_metadata}
\alias{grab_metadata}
\title{Grab Metadata}
\usage{
grab_metadata(x, samples = "SampleID", otus = "variable")
}
\arguments{
\item{samples}{The column header for your sample identifiers. Defaults to SampleID}

\item{otus}{The column header for your OTU identifiers. Defaults to variable}

\item{value}{The column header for the OTU abundances. Defaults to RA}

\item{dist}{The distance metric options from vegan's vegdist(). Defaults to bray}

\item{formula}{The formula to feed into adonis}
}
\value{
tibble of OTUs by Sample IDs
}
\description{
This function will take the long table and turn it back into a matrix with OTUs by Sample IDs.
}
\examples{
grab_metadata()
}
