% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_adonis.R
\name{long_adonis}
\alias{long_adonis}
\title{Adonis on Long Data}
\usage{
long_adonis(x, samples = "SampleID", otus = "variable", value = "RA",
  dist = "bray", formula)
}
\arguments{
\item{samples}{The column header for your sample identifiers. Defaults to SampleID}

\item{otus}{The column header for your OTU identifiers. Defaults to variable}

\item{value}{The column header for the OTU abundances. Defaults to RA}

\item{dist}{The distance metric options from vegan's vegdist(). Defaults to bray}

\item{formula}{The formula to feed into adonis}
}
\value{
S3 adonis object
}
\description{
This function is a wrapper around Vegan's adonis function for performing PERMANOVA on tidy (long) data.
}
\examples{
long_adonis()
}
\keyword{adonis}
