% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_functions.R
\name{read_manuscript}
\alias{read_manuscript}
\title{read_manuscript}
\usage{
read_manuscript(address, id, current)
}
\arguments{
\item{address}{string url or file location of knited html}

\item{id}{the id from a html tag}

\item{current}{the current version of the paragraph (optional)}
}
\value{

}
\description{
read_manuscript
}
\examples{

}
