% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{theme_mb}
\alias{theme_mb}
\title{theme_mb
Loads an object for customising plots to an MnB theme}
\usage{
theme_mb()
}
\value{
A ggplot2 theme object
}
\description{
theme_mb
Loads an object for customising plots to an MnB theme
}
\examples{
\dontrun{
original <- iris \%>\%
  ggplot(aes(x = Sepal.Length, y = Sepal.Width, col = Species)) +
  geom_smooth(method = "lm", se = F) +
  geom_jitter()
mnb_version <- original + theme_mnb()
}
}
