% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_betadisper.R
\name{tidy_betadisper}
\alias{tidy_betadisper}
\title{Beta dispersion on Tidy Data}
\usage{
tidy_betadisper(x, samples = "SampleID", otus = "variable", value = "RA",
  dist = "bray", group = NULL)
}
\arguments{
\item{samples}{The column header for your sample identifiers. Defaults to SampleID}

\item{otus}{The column header for your OTU identifiers. Defaults to variable}

\item{value}{The column header for the OTU abundances. Defaults to RA}

\item{dist}{The distance metric options from vegan's vegdist(). Defaults to bray}

\item{group}{A column in the table that identifies the independent variable}
}
\value{
A list containing the axes, the eigenvalues, and loadings (optional)
}
\description{
This function is a wrapper around Vegan's betadisper function for analyzing dispersion within groups on tidy (long) data.
}
\examples{
tidy_betadisper()
}
\keyword{PCoA}
