% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_cap.R
\name{tidy_cap}
\alias{tidy_cap}
\title{Constrained Ordination on Tidy Data}
\usage{
tidy_cap(x, samples = "SampleID", otus = "variable", value = "RA",
  dist = "bray", formula = NULL)
}
\arguments{
\item{samples}{The column header for your sample identifiers. Defaults to SampleID}

\item{otus}{The column header for your OTU identifiers. Defaults to variable}

\item{value}{The column header for the OTU abundances. Defaults to RA}

\item{dist}{The distance metric options from vegan's vegdist(). Defaults to bray}

\item{formula}{The formula to feed into the capscale function}
}
\value{
A list containing the axes, the eigenvalues, and loadings (optional)
}
\description{
This function is a wrapper around Vegan's capscale function for performing PCoA on tidy (long) data.
}
\examples{
tidy_cap()
}
\keyword{PCoA}
