% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pcoa.R
\name{tidy_pcoa}
\alias{tidy_pcoa}
\title{PCoA on Long Data}
\usage{
tidy_pcoa(x, samples = "SampleID", otus = "variable", value = "RA",
  dist = "bray")
}
\arguments{
\item{samples}{The column header for your sample identifiers. Defaults to SampleID}

\item{otus}{The column header for your OTU identifiers. Defaults to variable}

\item{value}{The column header for the OTU abundances. Defaults to RA}

\item{dist}{The distance metric options from vegan's vegdist(). Defaults to bray}
}
\value{
A list containing the axes and the eigenvalues
}
\description{
This function is a wrapper around Vegan's capscale function for performing PCoA on tidy (long) data.
}
\examples{
tidy_pcoa()
}
\keyword{PCoA}
