% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_rarefy.R
\name{tidy_rarefy}
\alias{tidy_rarefy}
\title{Rarefy Long Data}
\usage{
tidy_rarefy(x, rare_depth = NULL, samples = "SampleID", otus = "variable",
  value = "value")
}
\arguments{
\item{rare_depth}{The sampling depth for rarefaction. Defaults to the minimum sample depth if no value is provided}

\item{samples}{The column header for your sample identifiers. Defaults to SampleID}

\item{otus}{The column header for your OTU identifiers. Defaults to variable}

\item{value}{The column header for the OTU abundances. Defaults to RA}
}
\value{
A list containing the axes, the eigenvalues, and loadings (optional)
}
\description{
This function is a wrapper around Vegan's capscale function for performing PCoA on tidy (long) data.
}
\examples{
tidy_rarefy()
}
\keyword{Rarefaction}
