% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDGEL.R
\name{tidyDGEL}
\alias{tidyDGEL}
\title{Create a DGEL object from tidy table}
\usage{
tidyDGEL(x, method = "edgeR", samples = "SampleID", otus = "variable",
  value = "value", group)
}
\arguments{
\item{method}{The method for doing differential gene expression. Defaults to edgeR.}

\item{samples}{The column header for your sample identifiers. Defaults to SampleID.}

\item{otus}{The column header for your OTU identifiers. Defaults to variable.}

\item{value}{The column header for the OTU abundances. Defaults to RA.}

\item{group}{The variable that the user is trying to differentiate. Needs to be a column in the dataset.}
}
\value{
tibble of OTUs by Sample IDs
}
\description{
This function will take the long table and return a DGEL object either from edgeR or DESeq2. At the moment, it can only do simple designs (limited to one factor).
}
\examples{
widen()
}
