% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_mbtemplate.R
\name{use_tidytemplate}
\alias{use_tidytemplate}
\title{Call and open the M&B template}
\usage{
use_tidytemplate(name = NULL, open = interactive(), ...)
}
\arguments{
\item{name}{name of your paper (YYYY_DESCRIPTION_MS.Rmd)}

\item{open}{should the file be opened after being created}

\item{...}{arguments to be passed to \link[usethis]{use_template}}
}
\description{
Use the ttemplate Rmd for your paper
}
\examples{
\donttest{
if(interactive()){
  use_mbtemplate(name = "2020_iPLAY_MS.Rmd")
}
}

}
