% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_functions.R
\name{get_pdf_pagenumber}
\alias{get_pdf_pagenumber}
\title{get_page_number}
\usage{
get_pdf_pagenumber(string, pdf_path, max.distance = 0.15)
}
\arguments{
\item{string}{text to match}

\item{pdf_path}{file to search}

\item{max.distance}{argument passed to agrep}
}
\description{
Finds page number from pdf based on text matching
}
