% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revision_functions.R
\name{get_pdf_pagenumber}
\alias{get_pdf_pagenumber}
\title{get_page_number}
\usage{
get_pdf_pagenumber(string, path, max.distance = 0.4)
}
\arguments{
\item{string}{text to match}

\item{path}{file to search}

\item{max.distance}{scalar. The proportion that a match can vary from the input string (see agrep)}
}
\description{
Finds page number from pdf based on text matching
}
