% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_metadata.R
\name{grab_metadata}
\alias{grab_metadata}
\title{Grab Metadata}
\usage{
grab_metadata(x, samples = "SampleID", otus = "variable")
}
\arguments{
\item{samples}{The column header for your sample identifiers. Defaults to SampleID}

\item{otus}{The column header for your OTU identifiers. Defaults to variable}
}
\value{
tibble of OTUs by Sample IDs
}
\description{
This function will take a tidy dataframe and return the distinct information about each sample.
}
\examples{
grab_metadata()
}
