% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplay_pal.R
\name{print.palette}
\alias{print.palette}
\title{print.palette}
\usage{
\method{print}{palette}(x, ...)
}
\arguments{
\item{x}{palette a structure of class palette}

\item{...}{additional arguments passed to text}
}
\value{

}
\description{
print.palette
}
\examples{
iplay_pal()
}
